fix: 최초 로드 시 treeApi를 받아오지 못하는 문제 해결
