+++ frontend/techpick/src/app/page.tsx
@@ -12,13 +12,8 @@ import { NodeApi } from 'react-arborist';
 import { useTreeStore } from '@/shared/stores/treeStore';
 
 export default function MainPage() {
-  const {
-    focusedNode,
-    focusedFolderNodeList,
-    focusedLinkNodeList,
-    setFocusedFolderNodeList,
-    setFocusedLinkNodeList,
-  } = useTreeStore();
+  const { focusedNode, setFocusedFolderNodeList, setFocusedLinkNodeList } =
+    useTreeStore();
 
   const [tempFocusedFolderList, tempFocusedPickList] = useMemo(() => {
     if (!focusedNode || !focusedNode.children?.length) {
@@ -53,11 +48,7 @@ export default function MainPage() {
       <div className={viewWrapper}>
         <div className={viewContainer}>
           <DirectoryTreeSection />
-          <LinkEditorSection
-            focusedNode={focusedNode}
-            focusedNodeFolder={focusedFolderNodeList}
-            focusedNodeLink={focusedLinkNodeList}
-          />
+          <LinkEditorSection />
           <FeaturedSection />
         </div>
       </div>
